Made changes as per PR, created seperate modules for sql,firestore and bigtable, implemented transaction based approach
