+++ cartography/data/indexes.cypher
@@ -94,8 +94,12 @@ CREATE INDEX ON :GSuiteUser(id);
 CREATE INDEX ON :GCPFunction(id);
 CREATE INDEX ON :GCPSQLInstance(id);
 CREATE INDEX ON :GCPSQLUser(id);
-CREATE INDEX ON :GCPBigtable(id);
-CREATE INDEX ON :GCPFirestore(id);
+CREATE INDEX ON :GCPBigtableInstance(id);
+CREATE INDEX ON :GCPBigtableCluster(id);
+CREATE INDEX ON :GCPBigtableClusterBackup(id);
+CREATE INDEX ON :GCPBigtableTable(id);
+CREATE INDEX ON :GCPFirestoreDatabase(id);
+CREATE INDEX ON :GCPFirestoreIndex(id);
 CREATE INDEX ON :Ip(id);
 CREATE INDEX ON :Ip(ip);
 CREATE INDEX ON :IpPermissionInbound(ruleid);
