+++ cartography/intel/gcp/__init__.py
@@ -18,7 +18,9 @@ from cartography.intel.gcp import dns
 from cartography.intel.gcp import gke
 from cartography.intel.gcp import storage
 from cartography.intel.gcp import cloudfunction
-from cartography.intel.gcp import databases
+from cartography.intel.gcp import sql
+from cartography.intel.gcp import bigtable
+from cartography.intel.gcp import firestore
 from cartography.util import run_analysis_job
 from cartography.util import timeit
 
@@ -33,10 +35,10 @@ service_names = Services(
     storage='storage.googleapis.com',
     gke='container.googleapis.com',
     dns='dns.googleapis.com',
-    cloudfunction = 'cloudfunctions.googleapis.com',
-    cloudsql = 'sqladmin.googleapis.com',
-    cloudbigtable = 'bigtableadmin.googleapis.com',
-    firestore = 'firestore.googleapis.com'
+    cloudfunction='cloudfunctions.googleapis.com',
+    cloudsql='sqladmin.googleapis.com',
+    cloudbigtable='bigtableadmin.googleapis.com',
+    firestore='firestore.googleapis.com'
 )
 
 
@@ -116,7 +118,8 @@ def _get_serviceusage_resource(credentials: GoogleCredentials) -> Resource:
     """
     return googleapiclient.discovery.build('serviceusage', 'v1', credentials=credentials, cache_discovery=False)
 
-def _get_cloudfunction_resource(credentials:GoogleCredentials) -> Resource:
+
+def _get_cloudfunction_resource(credentials: GoogleCredentials) -> Resource:
     """
     Instantiates a cloud function resource object.
     See: https://cloud.google.com/functions/docs/reference/rest
@@ -126,7 +129,8 @@ def _get_cloudfunction_resource(credentials:GoogleCredentials) -> Resource:
     """
     return googleapiclient.discovery.build('cloudfunctions', 'v1', credentials=credentials, cache_discovery=False)
 
-def _get_cloudsql_resource(credentials:GoogleCredentials) -> Resource:
+
+def _get_cloudsql_resource(credentials: GoogleCredentials) -> Resource:
     """
     Instantiates a cloud sql resource object.
     See: https://cloud.google.com/sql/docs/mysql/admin-api/rest
@@ -134,7 +138,8 @@ def _get_cloudsql_resource(credentials:GoogleCredentials) -> Resource:
     :param credentials: The GoogleCredentials object
     :return: A serviceusage resource object
     """
-    return googleapiclient.discovery.build('sqladmin', 'v1', credentials=credentials,cache_discovery=False)
+    return googleapiclient.discovery.build('sqladmin', 'v1', credentials=credentials, cache_discovery=False)
+
 
 def _get_cloudbigtable_resource(credentials: GoogleCredentials) -> Resource:
     """
@@ -146,6 +151,7 @@ def _get_cloudbigtable_resource(credentials: GoogleCredentials) -> Resource:
     """
     return googleapiclient.discovery.build('bigtableadmin', 'v2', credentials=credentials, cache_discovery=False)
 
+
 def _get_firestore_resource(credentials: GoogleCredentials) -> Resource:
     """
     Instantiates a cloud firestore resource object.
@@ -154,6 +160,7 @@ def _get_firestore_resource(credentials: GoogleCredentials) -> Resource:
     :return: A serviceusage resource object
     """
     return googleapiclient.discovery.build('firestore', 'v1', credentials=credentials, cache_discovery=False)
+
 
 def _initialize_resources(credentials: GoogleCredentials) -> Resource:
     """
@@ -169,10 +176,10 @@ def _initialize_resources(credentials: GoogleCredentials) -> Resource:
         container=_get_container_resource(credentials),
         serviceusage=_get_serviceusage_resource(credentials),
         dns=_get_dns_resource(credentials),
-        cloudfunction = _get_cloudfunction_resource(credentials),
-        cloudsql = _get_cloudsql_resource(credentials),
-        cloudbigtable = _get_cloudbigtable_resource(credentials),
-        firestore = _get_firestore_resource(credentials),
+        cloudfunction=_get_cloudfunction_resource(credentials),
+        cloudsql=_get_cloudsql_resource(credentials),
+        cloudbigtable=_get_cloudbigtable_resource(credentials),
+        firestore=_get_firestore_resource(credentials),
     )
 
 
@@ -232,11 +239,11 @@ def _sync_single_project(
     if service_names.cloudfunction in enabled_services:
         cloudfunction.sync(neo4j_session, resources.cloudfunction, project_id, gcp_update_tag, common_job_parameters)
     if service_names.cloudsql in enabled_services:
-        databases.sync_sql(neo4j_session, resources.cloudsql, project_id,gcp_update_tag, common_job_parameters)
+        sql.sync_sql(neo4j_session, resources.cloudsql, project_id, gcp_update_tag, common_job_parameters)
     if service_names.cloudbigtable in enabled_services:
-        databases.sync_bigtable(neo4j_session, resources.cloudbigtable, project_id, gcp_update_tag, common_job_parameters)
+        bigtable.sync_bigtable(neo4j_session, resources.cloudbigtable, project_id, gcp_update_tag, common_job_parameters)
     if service_names.firestore in enabled_services:
-        databases.sync_firestore(neo4j_session, resources.firestore, project_id, gcp_update_tag, common_job_parameters)
+        firestore.sync_firestore(neo4j_session, resources.firestore, project_id, gcp_update_tag, common_job_parameters)
 
 
 def _sync_multiple_projects(
@@ -265,6 +272,7 @@ def _sync_multiple_projects(
         _sync_single_project(neo4j_session, resources, project_id, gcp_update_tag, common_job_parameters)
 
     del common_job_parameters["GCP_PROJECT_ID"]
+
 
 @timeit
 def start_gcp_ingestion(neo4j_session: neo4j.Session, config: Config) -> None:
